+++ packages/apptive_grid_form/test/apptive_grid_form_test.dart
@@ -346,12 +346,13 @@ void main() {
       });
 
       testWidgets('Back to Form shows Form', (tester) async {
+        final formUri = RedirectFormUri(
+          components: ['form'],
+        );
         final target = TestApp(
           client: client,
           child: ApptiveGridForm(
-            formUri: RedirectFormUri(
-              components: ['form'],
-            ),
+            formUri: formUri,
           ),
         );
         final action = FormAction('uri', 'method');
@@ -364,7 +365,7 @@ void main() {
         );
 
         when(
-          () => client.loadForm(formUri: RedirectFormUri(components: ['form'])),
+          () => client.loadForm(formUri: formUri),
         ).thenAnswer((realInvocation) async => formData);
         when(() => client.performAction(action, formData))
             .thenAnswer((_) => Future.error(''));
@@ -379,6 +380,8 @@ void main() {
         await tester.pumpAndSettle();
 
         expect(find.text('Form Title'), findsOneWidget);
+        // Don't reload here
+        verify(() => client.loadForm(formUri: formUri)).called(1);
       });
 
       testWidgets('Cache Response. Additional Answer shows Form',
@@ -429,6 +432,77 @@ void main() {
 
         expect(find.text('Form Title'), findsOneWidget);
         verify(() => client.loadForm(formUri: any(named: 'formUri'))).called(2);
+      });
+    });
+
+    group('Error Message', () {
+      testWidgets('Error shows to String Error', (tester) async {
+        final target = TestApp(
+          client: client,
+          child: ApptiveGridForm(
+            formUri: RedirectFormUri(
+              components: ['form'],
+            ),
+          ),
+        );
+        final action = FormAction('uri', 'method');
+        final formData = FormData(
+          name: 'Form Name',
+          title: 'Form Title',
+          components: [],
+          actions: [action],
+          schema: {},
+        );
+
+        when(
+          () => client.loadForm(formUri: RedirectFormUri(components: ['form'])),
+        ).thenAnswer((realInvocation) async => formData);
+        when(() => client.performAction(action, formData))
+            .thenAnswer((_) => Future.error(Exception('Testing Errors')));
+
+        await tester.pumpWidget(target);
+        await tester.pumpAndSettle();
+        await tester.tap(find.byType(ActionButton));
+        await tester.pumpAndSettle();
+        expect(
+          find.text('Exception: Testing Errors', skipOffstage: false),
+          findsOneWidget,
+        );
+      });
+
+      testWidgets('Response shows Status Code and Body', (tester) async {
+        final target = TestApp(
+          client: client,
+          child: ApptiveGridForm(
+            formUri: RedirectFormUri(
+              components: ['form'],
+            ),
+          ),
+        );
+        final action = FormAction('uri', 'method');
+        final formData = FormData(
+          name: 'Form Name',
+          title: 'Form Title',
+          components: [],
+          actions: [action],
+          schema: {},
+        );
+
+        when(
+          () => client.loadForm(formUri: RedirectFormUri(components: ['form'])),
+        ).thenAnswer((realInvocation) async => formData);
+        when(() => client.performAction(action, formData)).thenAnswer(
+          (_) => Future.error(http.Response('Testing Errors', 400)),
+        );
+
+        await tester.pumpWidget(target);
+        await tester.pumpAndSettle();
+        await tester.tap(find.byType(ActionButton));
+        await tester.pumpAndSettle();
+        expect(
+          find.text('400: Testing Errors', skipOffstage: false),
+          findsOneWidget,
+        );
       });
     });
   });